@@ -9,24 +9,24 @@
 
 import { test } from '@japa/runner'
 import dedent from 'dedent'
-import { Tokenizer } from '../src/tokenizer'
-import { MustacheTypes } from '../src/types'
+import { Tokenizer } from '../src/tokenizer.js'
+import { MustacheTypes } from '../src/types.js'
 
 const tagsDef = {
   if: class If {
-    public static block = true
-    public static selfclosed = false
-    public static seekable = true
+    static block = true
+    static selfclosed = false
+    static seekable = true
   },
   else: class Else {
-    public static block = false
-    public static selfclosed = false
-    public static seekable = false
+    static block = false
+    static selfclosed = false
+    static seekable = false
   },
   include: class Include {
-    public static block = false
-    public static selfclosed = false
-    public static seekable = true
+    static block = false
+    static selfclosed = false
+    static seekable = true
   },
 }
 
